WIP : Add third_party support of weston to chromium.
This is the first version of the patch. In this patch the config and sources
are propoerly generated for weston. Only the problem is, for fully
operational it will be needed to support wayland and wayland-protocols
suppport in third_party of chromium as well.

(cherry picked from commit ad5382201a545cd3343f49e147957c8e47cf8dfe)

Bug: Add third_party weston support in chromium.
Change-Id: I5ef4e90c312df65078cf7caa89b64f672f349b60